pyproject.toml
[project]
name = "image-generator"
version = "0.1.0"
description = "A MCP server project for generating and saving images using Replicate"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.3.0",
    "replicate>=0.22.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "requests>=2.31.0",
]
[[project.authors]]
name = "rmcendarfer2017"
email = "ryanmcendarfer@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
image-generator = "image_generator:main"
